Add bottom panel component to DI
