+++ data/__init__.py
@@ -1,8 +1,12 @@
 from .resolver import DataResolver, StaticDataResolver
 from .purpleair import PurpleAirDataResolver
+from .envcanada import EnvironmentCanadaDataResolver
+from .calendar import CalendarDataResolver
 
 __all__ = [
+    'CalendarDataResolver',
     'DataResolver',
+    'EnvironmentCanadaDataResolver',
     'PurpleAirDataResolver',
     'StaticDataResolver',
 ]
